Add public header
